refactor(time): 删除TimeMode枚举，改为连续速度倍数+暂停状态
- WorldClock: tick()不再接收dt，fast_forward改为run_to，set_mode改为直接设置speed/paused
- 终端执行器适配新的速度API，支持任意数值速度
- world_tree测试timestamp从float改为int
- i18n文本从模式概念更新为速度概念（英文/中文）
- 更新game_loop、net、console等测试

+++ ascend-backend/ascend/entity/entity.py
@@ -41,7 +41,7 @@ class Entity:
     chunk_y: int
     tile_x: int | None
     tile_y: int | None
-    spawned_at: float
+    spawned_at: int
 
     id: str = field(default_factory=lambda: uuid.uuid4().hex)
     data: dict | None = None
